requirements.txt
requests~=2.31.0
Pillow~=9.5.0
selenium~=4.10.0
opencv-python~=4.7.0.72
numpy~=1.25.0
scikit-image~=0.21.0